+++ src/api/parcel/controllers/2.0.0/parcels.controller.js
@@ -1,6 +1,5 @@
 import Boom from '@hapi/boom'
 import { statusCodes } from '~/src/api/common/constants/status-codes.js'
-import { sqmToHaRounded } from '~/src/api/common/helpers/measurement.js'
 import {
   errorResponseSchema,
   internalServerErrorResponseSchema
@@ -8,27 +7,27 @@ import {
 import {
   parcelsSchema,
   parcelsSuccessResponseSchema
-} from '~/src/api/parcel/schema/parcel.schema.js'
-import { getParcelActionsWithAvailableArea } from '~/src/api/parcel/service/parcel.service.js'
-import { sizeTransformer } from '~/src/api/parcel/transformers/parcelActions.transformer.js'
-import { getAgreementsForParcel } from '~/src/api/agreements/queries/getAgreementsForParcel.query.js'
-import { mergeAgreementsTransformer } from '~/src/api/agreements/transformers/agreements.transformer.js'
-import { getDataAndValidateRequest } from '../validation/parcel.validation.js'
+} from '~/src/api/parcel/schema/2.0.0/parcel.schema.js'
+import { getDataAndValidateRequest } from '../../validation/2.0.0/parcel.validation.js'
 import { createCompatibilityMatrix } from '~/src/available-area/compatibilityMatrix.js'
 import {
   logBusinessError,
   logInfo,
   logValidationWarn
 } from '~/src/api/common/helpers/logging/log-helpers.js'
+import {
+  getActionsForParcel,
+  getActionsForParcelWithSSSIConsentRequired
+} from '../../service/parcel.service.js'
 
 /**
  * ParcelsController
  * Returns a single land parcel merged with land actions
  * @satisfies {Partial<ServerRoute>}
  */
-const ParcelsController = {
+const ParcelsControllerV2 = {
   options: {
-    tags: ['api'],
+    tags: ['api/v2'],
     description: 'Get multiple land parcels with selected fields',
     notes:
       'Returns data for multiple parcels and includes the requested fields',
@@ -102,10 +101,26 @@ const ParcelsController = {
         })
       )
 
+      let transformedResponseParcels = responseParcels
+
+      if (
+        parcelIds.length === 1 &&
+        fields.includes('actions.sssiConsentRequired')
+      ) {
+        transformedResponseParcels =
+          await getActionsForParcelWithSSSIConsentRequired(
+            parcelIds,
+            responseParcels,
+            validationResponse.enabledActions,
+            request.logger,
+            postgresDb
+          )
+      }
+
       return h
         .response({
           message: 'success',
-          parcels: responseParcels
+          parcels: transformedResponseParcels
         })
         .code(statusCodes.ok)
     } catch (error) {
@@ -124,56 +139,8 @@ const ParcelsController = {
     }
   }
 }
-export { ParcelsController }
+export { ParcelsControllerV2 }
 
 /**
  * @import { ServerRoute } from '@hapi/hapi'
  */
-
-async function getActionsForParcel(
-  parcel,
-  payload,
-  showActionResults,
-  enabledActions,
-  compatibilityCheckFn,
-  request
-) {
-  const { fields, plannedActions } = payload
-
-  const parcelResponse = {
-    parcelId: parcel.parcel_id,
-    sheetId: parcel.sheet_id
-  }
-
-  if (fields.includes('size')) {
-    parcelResponse.size = sizeTransformer(sqmToHaRounded(parcel.area_sqm))
-  }
-
-  if (fields.some((f) => f.startsWith('actions'))) {
-    const agreements = await getAgreementsForParcel(
-      parcel.sheet_id,
-      parcel.parcel_id,
-      request.server.postgresDb,
-      request.logger
-    )
-
-    const mergedActions = mergeAgreementsTransformer(agreements, plannedActions)
-
-    const actionsWithAvailableArea = await getParcelActionsWithAvailableArea(
-      parcel,
-      mergedActions,
-      showActionResults,
-      enabledActions,
-      compatibilityCheckFn,
-      request.server.postgresDb,
-      request.logger
-    )
-
-    const sortedParcelActions = actionsWithAvailableArea.toSorted((a, b) =>
-      a.code.localeCompare(b.code)
-    )
-
-    parcelResponse.actions = sortedParcelActions
-  }
-  return parcelResponse
-}