+++ src/api/parcel/service/parcel.service.js
@@ -4,8 +4,16 @@ import {
 } from '~/src/available-area/availableArea.js'
 import {
   actionTransformer,
-  plannedActionsTransformer
+  plannedActionsTransformer,
+  sizeTransformer,
+  sssiConsentRequiredActionTransformer
 } from '~/src/api/parcel/transformers/parcelActions.transformer.js'
+import { sqmToHaRounded } from '~/src/api/common/helpers/measurement.js'
+import { getAgreementsForParcel } from '~/src/api/agreements/queries/getAgreementsForParcel.query.js'
+import { mergeAgreementsTransformer } from '~/src/api/agreements/transformers/agreements.transformer.js'
+import { getDataLayerQuery } from '~/src/api/data-layers/queries/getDataLayer.query.js'
+import { executeSingleRuleForEnabledActions } from '~/src/rules-engine/rulesEngine.js'
+import { sssiConsentRequired } from '~/src/rules-engine/rules/1.0.0/sssi-consent-required.js'
 
 /**
  * @import {LandParcelDb} from '~/src/api/parcel/parcel.d.js'
@@ -94,4 +102,93 @@ export async function getParcelActionsWithAvailableArea(
   }
 
   return actionsWithAvailableArea
+}
+
+export async function getActionsForParcel(
+  parcel,
+  payload,
+  showActionResults,
+  enabledActions,
+  compatibilityCheckFn,
+  request
+) {
+  const { fields, plannedActions } = payload
+
+  const parcelResponse = {
+    parcelId: parcel.parcel_id,
+    sheetId: parcel.sheet_id
+  }
+
+  if (fields.includes('size')) {
+    parcelResponse.size = sizeTransformer(sqmToHaRounded(parcel.area_sqm))
+  }
+
+  if (fields.some((f) => f.startsWith('actions'))) {
+    const agreements = await getAgreementsForParcel(
+      parcel.sheet_id,
+      parcel.parcel_id,
+      request.server.postgresDb,
+      request.logger
+    )
+
+    const mergedActions = mergeAgreementsTransformer(agreements, plannedActions)
+
+    const actionsWithAvailableArea = await getParcelActionsWithAvailableArea(
+      parcel,
+      mergedActions,
+      showActionResults,
+      enabledActions,
+      compatibilityCheckFn,
+      request.server.postgresDb,
+      request.logger
+    )
+
+    const sortedParcelActions = actionsWithAvailableArea.toSorted((a, b) =>
+      a.code.localeCompare(b.code)
+    )
+
+    parcelResponse.actions = sortedParcelActions
+  }
+  return parcelResponse
+}
+
+export async function getActionsForParcelWithSSSIConsentRequired(
+  parcelIds,
+  responseParcels,
+  enabledActions,
+  logger,
+  postgresDb
+) {
+  const { sheetId, parcelId } = splitParcelId(parcelIds[0], logger)
+
+  const intersectingAreaPercentage = await getDataLayerQuery(
+    sheetId,
+    parcelId,
+    postgresDb,
+    logger
+  )
+
+  const application = {
+    areaAppliedFor: 0,
+    actionCodeAppliedFor: '',
+    landParcel: {
+      area: 0,
+      existingAgreements: [],
+      intersections: {
+        sssi: { intersectingAreaPercentage }
+      }
+    }
+  }
+
+  const sssiConsentRequiredAction = executeSingleRuleForEnabledActions(
+    enabledActions,
+    application,
+    'sssi-consent-required',
+    sssiConsentRequired
+  )
+
+  return sssiConsentRequiredActionTransformer(
+    responseParcels,
+    sssiConsentRequiredAction
+  )
 }